Perfilado Main, metodo addParque, nada testeado por falta de BD.

+++ ud8_boletin_ej1/src/main/java/com/pablo/ud8_boletin_ej1/DAO/Ciudad/ICiudadDAO.java
@@ -16,7 +16,7 @@ public interface ICiudadDAO {
 
    public int add(Ciudad ciudad);
 
-   public int cantidadParquesPorCiudad(Ciudad ciudad);
+   public long cantidadParquesPorCiudad(Ciudad ciudad);
 
    public List<Ciudad> ciudadesPorTamanyoParque(Parque parque);
 }
